issue #2 moved classic fileInfos to their own files
