+++ lib/engines/classic/classic-file-info.js
@@ -1,0 +1,11 @@
+var FileInfo = require('../../models/file-info');
+
+var PLURAL_OVERRIDES = {
+  'mirage': 'mirage',
+  'html': 'html'
+};
+var ClassicFileInfo = FileInfo.extend({
+  plural_overrides: PLURAL_OVERRIDES
+});
+
+module.exports = ClassicFileInfo;
