+++ lib/engines/classic/single-type-collection-file-info.js
@@ -1,0 +1,15 @@
+var ClassicFileInfo = require('./classic-file-info');
+
+var SingleTypeCollectionFileInfo = ClassicFileInfo.extend({
+  populateDestination: function() {
+    if (this.destRelativePath) {
+      return this.destRelativePath;
+    }
+
+    var destRelativePath = this.collection + '/' + this.name + this.ext;
+
+    return this.destRelativePath = destRelativePath;
+  }
+});
+
+module.exports = SingleTypeCollectionFileInfo;
