+++ lib/utils/type-included.js
@@ -1,0 +1,24 @@
+var typeIncluded = function(type) {
+  switch (type) {
+    // data
+  case 'adapter':
+  case 'serializer':
+  case 'model':
+    return false;
+
+    // routes
+  case 'route':
+  case 'controller':
+  case 'template':
+    return false;
+
+    // components
+  case 'component':
+    return false;
+
+  default:
+    return true;
+  }
+};
+
+module.exports = typeIncluded;
